Rename: common의 하위 폴더Navbar, Modal 생성 및 파일분리(#65)
- components/common의 파일이 많아져서 Modal / Navbar 폴더 생성하고 각 파일들 폴더 안으로 이동
