+++ client/src/components/Common/Modal/FriendInfo.module.scss
