+++ client/src/components/Common/Modal/FriendInfo.tsx
