+++ client/src/components/Common/Modal/Modal.module.scss
@@ -1,6 +1,6 @@
-@use '../../styles/color' as C;
-@use '../../styles/typography' as T;
-@use '../../styles/utils' as U;
+@use '../../../styles/color' as C;
+@use '../../../styles/typography' as T;
+@use '../../../styles/utils' as U;
 
 .background {
   position: fixed;
