+++ client/src/components/Common/Navbar/NavbarProfile.module.scss
@@ -1,6 +1,6 @@
-@use '../../styles/utils' as U;
-@use '../../styles/typography' as T;
-@use '../../styles/color' as C;
+@use '../../../styles/utils' as U;
+@use '../../../styles/typography' as T;
+@use '../../../styles/color' as C;
 
 .profileBox {
   @include U.size(120px);
